requirements.txt
requests
json
discord.py
ta
ccxt
hashlib